+++ tests/testthat/test_c14_date_list_country_attribution.R
@@ -2,21 +2,21 @@ context("country attribution functions")
 
 #### output ####
 
-# standardize_country_name
+# fix_database_country_name
 
-result <- standardize_country_name(
+result <- fix_database_country_name(
   example_c14_date_list,
   quiet = TRUE
 )
 
-test_that("standardize_country_name gives back a c14_date_list", {
+test_that("fix_database_country_name gives back a c14_date_list", {
   expect_s3_class(
     result,
     "c14_date_list"
   )
 })
 
-test_that("standardize_country_name gives back a c14_date_list with the additional
+test_that("fix_database_country_name gives back a c14_date_list with the additional
           column country_thes", {
   expect_true(
     all(
@@ -26,7 +26,7 @@ test_that("standardize_country_name gives back a c14_date_list with the addition
   )
 })
 
-test_that("standardize_country_name gives back a c14_date_list with the additional
+test_that("fix_database_country_name gives back a c14_date_list with the additional
           column country_thes and this column is of type character", {
   expect_type(
     result$country_thes,
@@ -65,9 +65,9 @@ test_that("determine_country_by_coordinate gives back a c14_date_list with the a
 
 #### messages ####
 
-test_that("standardize_country_name has message output, if quiet == FALSE", {
+test_that("fix_database_country_name has message output, if quiet == FALSE", {
   expect_message(
-    standardize_country_name(
+    fix_database_country_name(
       example_c14_date_list,
       quiet = FALSE
     ),